+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # upcoming
 - pastes are now public by default
 - latest public pastes are now shown on the main page
+- `YAP_PASTE_RATE_LIMIT` now controls how often new pastes can be created from the same IP
 
 # 0.2.0
 - banner is now a link